rename table-component to list-component
