+++ nexuse2e-webapp/src/main/frontend/src/app/list/list.component.spec.ts
@@ -1,14 +1,14 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { TableComponent } from "./table.component";
+import { ListComponent } from "./list.component";
 import { MatTableModule } from "@angular/material/table";
 import { MatCheckboxModule } from "@angular/material/checkbox";
 import { RouterTestingModule } from "@angular/router/testing";
 import { TranslateModule } from "@ngx-translate/core";
 
 describe("TableComponent", () => {
-  let component: TableComponent;
-  let fixture: ComponentFixture<TableComponent>;
+  let component: ListComponent;
+  let fixture: ComponentFixture<ListComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -18,12 +18,12 @@ describe("TableComponent", () => {
         RouterTestingModule,
         TranslateModule.forRoot(),
       ],
-      declarations: [TableComponent],
+      declarations: [ListComponent],
     }).compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(TableComponent);
+    fixture = TestBed.createComponent(ListComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
